tests: added test for jelly integration (PR #99) - adding requested changes
