+++ test/jelly/parse_inputs_acc_ttl/test_parse_inputs_acc_ttl.py
@@ -1,0 +1,35 @@
+"""
+Tests for yatter jelly integration for input output.
+For PR #99: feat(cli): support for Jelly format for input/output
+"""
+
+import pytest
+import types
+from pathlib import Path
+import os
+
+import rdflib
+
+from yatter import translate
+
+pytest.importorskip("pyjelly")
+from yatter.cli import parse_inputs, write_results
+from rdflib import Graph
+from rdflib import compare
+
+def _mimic_args(**kwargs):
+    return types.SimpleNamespace(**kwargs)
+
+
+def test_parse_inputs_acc_ttl(tmp_path: Path):
+    output_mapping_path = str(tmp_path / 'output_mapping.yml')
+
+    args = _mimic_args(input_mapping_path = str(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'jelly_mapping_test_ttl.ttl')),
+                       output_mapping_path = output_mapping_path,
+                       merge_mapping = None,
+                       format=None)
+
+    parsed_format, parsed_data = parse_inputs(args)
+
+    assert isinstance(parsed_data, rdflib.Graph)
+    assert isinstance(parsed_format, str)
