+++ test/jelly/write_results_roundtrip_to_ttl/test_write_results_roundtrip_to_ttl.py
@@ -1,0 +1,40 @@
+"""
+Tests for yatter jelly integration for input output.
+For PR #99: feat(cli): support for Jelly format for input/output
+"""
+from unittest import expectedFailure
+
+import pytest
+import types
+from pathlib import Path
+import os
+
+import rdflib
+pytest.importorskip("pyjelly")
+from yatter.cli import parse_inputs, translate,write_results
+from rdflib import Graph
+from rdflib import compare
+
+def _mimic_args(**kwargs):
+    return types.SimpleNamespace(**kwargs)
+
+def test_write_results_roundtrip_to_ttl(tmp_path: Path):
+    output_mapping_path = str(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'output_mapping.ttl'))
+    args = _mimic_args(input_mapping_path=str(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'jelly_mapping_test_yml.yml')),
+                       output_mapping_path=output_mapping_path, merge_mapping=None,
+                       format=None)
+
+    parsed_format, parsed_data = parse_inputs(args)
+    mapping_content = translate(parsed_data, parsed_format)
+    assert isinstance(mapping_content, str)
+    write_results(args, mapping_content)
+
+    translated_mapping = Graph()
+    translated_mapping.parse(output_mapping_path, format='turtle')
+
+    expected_mapping = Graph()
+    expected_mapping.parse(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'jelly_mapping_test_ttl.ttl'), format='turtle')
+
+    assert isinstance(translated_mapping, rdflib.Graph)
+    assert compare.isomorphic(translated_mapping, expected_mapping)
+
